Update Corpus
Change the way Corpus creates the new Dictionary in Corpus :: process
Concretely, the Dictionary is queried with get only in case the entries
are needed making the implementation faster.

Change the way Corpus creates new WordSequences in
Corpus :: createWordSequences and Corpus :: processNumbers
Concretely, for the time being, no number is processed as a date until
a new way of number processing is added. Also, better all non-visible
characters are removed from the document text.