Refactor fetchers functionality
Folder doesn't contain a list of Emails but extends ArrayList.

EmailFetcher is not a Runnable. Implementation is left to the extending
classes. Also supports a default fetch method which will fetch the sent
mail and another fetch method for fetching any folder. It also provides
a list with all fetchable folders.

GmailFetcher implementation adapts to the changes to EmailFetcher.
Synchronization and thread safety is added.