Handle the streams of native processes
It seems that, if a java.lang.Process produces enough output or error
messages to the output or error stream, and fills the corresponding
buffer, the java.lang.Process will block until these messages are read.
To handle that, the output and error streams of LanguageModelProcess and
AcousticModelProcess are specifically redirected to files on the local
file system.