Show the number of Emails in each folder
On EmailTreeView, each folder knowns the number of the Emails it
contains and gets a state based on that number. Folders that do not
contain any Emails cannot be fetched while others can be fetched only
once. Selection is not allowed during fetching for the folder being
fetched.

The folder tree is created once, at the initialization of the
EmailTreeView. This means that, new folders, created after the
initialization of the EmailTreeView will not be fetched unless the
EmailListScene is reloaded.