Remove static from ImageViews in RecordController
According to the docs for Labeled :: setGraphic method:

The node specified for this variable cannot appear elsewhere in the
scene graph, otherwise the IllegalArgumentException is thrown.

Seems safer to keep a private version of each ImageView so that each
RecordController object uses its own versions.

A problem like this was faced in FolderValue so the ImageViews were
changed to private there also.