Fetch N most recent Emails
EmailTreeView is changed to EmailTreePane which will also contain
a Label and an IntegerField. The value of the IntegerField will be
used by the user to choose the number of the most recent Emails that
should be fetched in each directory.

FolderValue will use a HashEmailFetcher instead of a real EmailFetcher.
This is done to avoid passing the IntegerField to the FolderValue.
Through the HasEmailFetcher, a FolderValue can request to be fetched
and the EmailTreePane will provide the count of the Emails to be
fetched.

Make some correction in SortedMapEntryList method naming.

GMailFetcher now returns the N most recent Emails ordered with the
most recent being first and the least recent being last.