Fix bug in Dictionary and add documentation
On a CMU Sphinx dictionary, word entries start counting from 2. For
example:

the DH AH
the(2) DH IY

Up until now, the Dictionary implementation made the assumption that
the counting started from 1. This resulted in not finding any other
entry except the first one in getEntriesByKey method, not removing
any other entry except the first one in remove method and adding
entries with count equal to 1 in put method.