Fix bug in EmailTreePane
When a top folder was added on the EmailTreePane, the assumption that
the same folder handn't been added before was made. This resulted in
the following problem:

For folder structure "JetBrains/Intellij idea" the JavaMail API would
return the folders "JetBrains" and "JetBrains/Intellij idea". This
resulted in creating the folder "JetBrains" twice. The same problem
was encoutered with the "[Gmail]" folder and it was tackled by checking
the folder name with the notUsableFolder method inside GmailFetcher.

To solve this, the EmailTreePane will check if the top folder already
exists and if it does, it will simply copy the Value of the new folder
to the old one (since the new Value might contain better information
regarding the number of emails inside the folder, e.g. if
"JetBrains/Intellij idea" gets created first, there is no information
regarding the number of e-mails inside "JetBrains", so when "JetBrains"
arrives, its value should replace the old folder value which was
created as a parent for the "Intellij idea" folder with 0 e-mails).
Note that there is no need to check if the new folder contains e-mails
in order to copy them as children to the old one since, at this point,
only folders are created and e-mails will start arriving after the user
downloades a folder.

Also, the notUsableFolder check inside GmailFetcher is removed since it
is not needed anymore.